Move system-header-prefix to per flag suppressions
Remove third_party/gen path from system-header-prefix.
Use per flag suppressions in warning_suppression_mappings.txt
instead.

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/325155
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/tint/lang/wgsl/resolver/eval_test.h
@@ -141,7 +141,8 @@ inline void CheckConstant(const constant::Value* got_constant,
                         }
                         if (flags.float_compare) {
                             if (flags.float_compare_epsilon) {
-                                EXPECT_NEAR(got, expected, *flags.float_compare_epsilon)
+                                EXPECT_NEAR(static_cast<double>(got), static_cast<double>(expected),
+                                            static_cast<double>(*flags.float_compare_epsilon))
                                     << "index: " << i;
                             } else {
                                 EXPECT_FLOAT_EQ(got, expected) << "index: " << i;
